+++ src/components/Board/index.ts
@@ -1,1 +1,0 @@
-export { Board as default } from "./Board";