+++ src/components/Dashboard/index.ts
@@ -1,1 +1,4 @@
 export { Dashboard as default } from "./Dashboard";
+export * from "./DashboardCallout";
+export * from "./DashboardTheme";
+export * from "./DashboardWidget";