+++ CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @fluentui/react-teams
 
-This log was last generated on Thu, 19 Nov 2020 21:09:50 GMT and should not be manually modified.
+This log was last generated on Mon, 14 Dec 2020 19:57:06 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.1.0-alpha.10
+
+Mon, 14 Dec 2020 19:57:06 GMT
+
+### Changes
+
+- Board height fix (a.katrukhin@gmail.com)
+
+## 0.1.0-alpha.9
+
+Mon, 14 Dec 2020 19:20:03 GMT
+
+### Changes
+
+- Scroll issues fix (a.katrukhin@gmail.com)
+
+## 0.1.0-alpha.8
+
+Fri, 11 Dec 2020 18:35:34 GMT
+
+### Changes
+
+- Communication and Dashboard components are added, styles updates, Teams themes support improvements (a.katrukhin@gmail.com)
 
 ## 0.1.0-alpha.4